+++ kabuki/__init__.py
@@ -1,4 +1,4 @@
 from . import dataset
-from kabuki.hosting.hosting import upload_dataset, retrieve_dataset, list_datasets
+from kabuki.storage.hosting import upload_dataset, retrieve_dataset, list_datasets
 
 __version__ = "0.0.1"
