+++ kabuki/storage/hosting.py
